chore(kuma-http-api): expose/export openapi.yaml (#4334)
Allows us to import `openapi.yaml` which we can then use with our yaml
loader.

This could be temporary, later we might want to prebuild a json/js file
and export that instead. This would mean you wouldn't need a yaml loader
to be able to make use of this easily. For now tho, we have a yaml
loader already so this is fine for now.

Signed-off-by: John Cowen <john.cowen@konghq.com>