chore: centralize icons with XTheme and v-icon-start (#4284)
We have several areas where we use CSS based SVG icons, along with a
very "single use-case component" to add a component.

This PR does several things to centralize this sort of decorative
imagery:

- Adds `v-icon-start` and `v-icon-end` Vue directives for in template
declarative CSS icons using inline HTML/Vue.
- These directives are backed by a pure CSS approach to these sorts of
icons. It is possible to use `style="--icon-start"` or `element {
--icon-start: var(...) }` in pure CSS, but this should be an
escape-hatch/last resort if for whatever reason the Vue directive isn't
feasible for your use-case (reasons should be added to PRs)
- The CSS is backed by a new `XTheme` component which makes use of
scoped styling to apply styles only to children of `XTheme` i.e.
providing a centralized place to add icons and rules for those icons
without resorting to global styles.

We still have two "ways" to add icons but this is better than several
different implementations dotted around the codebase:

1. `v-icon-start` and `v-icon-end` for our own CSS based decorative
iconography
2. `XIcon` for when an inline "DOM element icon" is required

Currently `XIcon` also makes use of the new directives behind the
scenes, further centralizing, reducing duplication and promoting reuse.

---

`XTheme` will also be used in future to specify colorings and other
top-level "theming" and whilst we don't plan on adding dark mode support
in the immediate future, it will allow us to do so in a
centralized/non-global and composable way if we ever want to do that.

Signed-off-by: John Cowen <john.cowen@konghq.com>