chore(policies): refactor policy types to use more OpenAPI (#4333)
This PR started out as something else, but I eventually removed that.
The things I did on the way I figure are still a useful refactor, so
I've kept the PR with only the refactor.

Note: I think I'd like to start referring to the OpenAPI generated types
as `Kuma*` i.e. `KumaPolicy` instead of `GeneratedPolicy` like we have
been doing. This is a naming thing but reasons:

- Originally when we started using the word `Generated` it was to avoid
calling them `OpenApi*` or `Protobuf*`, the fact is they are `Generated`
somehow.
- I think `Kuma` is even better because its not important to use that
they are `Generated`. What is important is that they are the original
ones form `Kuma` itself - hence `Kuma*`

Anyway not a biggie and more of a thing that if we are amending the data
layer stuff I would propose that we default to using this `Kuma*` naming
to say "these are the ones from Kuma itself"

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>