chore(data): create @kumahq/data (#4335)
Does the bare minimum to move the majority of our `Data*` components and
related into their own `@kumahq/data`.

There are loads of little bits I would like to do here still, but I'd
rather merge this as is and then pick of all these bits as time goes on.
There's likely to be quite a few little PRs at some point in the future
to do this, but for now this at least creates the new package.

---

With the current work that I'm doing I keep bumping into places where
I'm like "oh I wish I'd already built `<thing>`, or I wish I'd cleaned
up DataSource so I can do `<thing>`".

I decided to bite the bullet with this and the first step is to create
our planned `@kumahq/data` package

Signed-off-by: John Cowen <john.cowen@konghq.com>