chore: adds new KumaTargetRef component (#4359)
New KumaTargetRef component for rendering a TargetRef consistently in
several different places.

I had to amend XDl slightly as its possible to have a single `dt` (in
the case of `Mesh` as opposed to `MeshService: name`)

---

Note: this may be tweaked slightly in the future, but for now it does
what I need it to do, including having a slot embedded incase we need to
put anything at the end of the badge.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>