feat(data-planes): traffic stats alerting (#4343)
Checks the stats for abnormalities and reports if abnormalities are
detected. For now we only check for certain `circuit_breakers` and
`outlier_detection` stats.

~I've added the detection to the parsing of the stats to match raw
values with certain regexes. If a pattern is detected it adds a warning
with the stat as a key and the reports as a value.~ The monitoring
happens during the nomalization of the stats after parsing. I've added a
new field `$meta.alerts.reports` to each `clusters` and `listeners`. The
`reports` is added as regex filtering separated with `|` to the query
when the user clicks on an inbound/outbound that is highlighted.

Closes https://github.com/kumahq/kuma-gui/issues/4325

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>