feat(x/notification): integrate toasts (#4360)
Integrates [Kongponent's
Toaster](https://kongponents.konghq.com/components/toaster.html) into
`XNotification` to provide a single component with different types to
either show a `banner` or `toast` notification. The main difference is
the appearance but also `toast`s are stateless and nothing is stored,
i.e. dismissals in local storage. The usage of `toast` is very similar
to `banner`. Unfortunately the API for `title` and `message` of
`Toaster` is limited to `string` only and the intended usage/triggering
is imperatively. To align both `banner` and `toast` usage I've used some
hacky way to use a default slot and kind of render it to a string, which
is then added to `toaster.open({ ... })`. Still we have to make sure to
only use `string` in the default slot as any HTML element is escaped and
rendered as a string.

```vue
<DataSink
  src="/foo"
  @change="(val) => {...}"
  v-slot="{ submit, writing, data }"
>
  <XNotification
    :notify="!writing && !!data"
    uri="foo-success"
    type="toast"
    variant="info"
  >
    {{ t('message.to.let.users.know.the.action.was.successful') }}
  </XNotification>
  <XAction appearance="primary" @click="() => submit({})">
    Click me
  </XAction>
</DataSink>
```

---

Some things we might want to change/improve in the future:

- The default `variant` of a notification is `warning` as this is the
default of a `banner`. The default `variant` of `toast` is rather
`info`. This means for now we should always apply `variant="info"` to
`type="toast"`.
- Since a `toast` is not tracked or stored we could remove the need for
the `uri`-prop.

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>