chore: allow DataCollection:EmptyState to be injectable (#4374)
Does some plumbing work to extract an dependency on `@kumahq/x`
(XEmptyState) from `DataCollection`.

This means we can totally amend the empty state from The Outside, say if
we need to add certain buttons if a list result is empty.

---

Separately from this I'd also like to do similar to our usage of
`KPagination` inside of `DataCollection` thus removing its dependency on
kongponents also, at which point I'll look to move DataCollection into
`@kumahq/data` where it belongs.

For the moment the work I'm doing depends on adding buttons to empty
states only so I only went so far in this PR.

Signed-off-by: John Cowen <john.cowen@konghq.com>