feat(oas): add oneOf implementation to MeshAccessLog (#4386)
This is a bit of a strange PR in that we can't move to "owning" the
entire schema for our policies, but I do what one policy here for
experimentation and illustration of what sort of thing we need.

This PR specifically adds a `oneOf` schema to `MeshAccessLog.backends`

Additional I've marked some fields as `readOnly`.

---

Just to underline again, this config is temporary, please ping me
offline if there are questions/concerns around this

Signed-off-by: John Cowen <john.cowen@konghq.com>