fix(dataplanes): don't show tp info for builtin gateway (#4393)
Closes https://github.com/kumahq/kuma-gui/issues/4356

Disables message for builtin gateways only.

The conditional change here is a bit weird as I wanted to ge the
wording/functionality right.

Built in gateways can't use transparent proxy so I added that to the
feature conditional i.e. `can('use transparent-proxy') === false`, but
in this case we don't want to show a message for built-in gateways, so I
added the conditional in two places to reflect the intention here.

Signed-off-by: John Cowen <john.cowen@konghq.com>