feat(data-planes/list): show cert info in case of SPIRE (#4400)
In case of SPIRE the certificate is managed externally, which is
indicated by existing `dataplaneInsight.mTLS` but without any expiration
times (only `certificateRegenerations` and `issuedBackend` which is a
KRI of a MeshIdentity). In case we can't show an
`certificateExpirationTime` but `mTLS` is defined we now show
`Externally managed`. Only in case `mTLS` is undefined we show `No
certificate`.

Closes https://github.com/kumahq/kuma-gui/issues/4384

Signed-off-by: schogges <moritz.fleck@konghq.com>