chore(renovate): extend kong's public-shared-renovate/javascript (#4416)
To benefit from the efforts put into collecting best practices for
renovate configs for the npm ecosystem, we've decided to extend Kong's
[public-shared-renovate/javascript base
config](https://github.com/Kong/public-shared-renovate/blob/main/base/javascript.json).

Closes https://github.com/kumahq/kuma-gui/issues/4410

After this is merged we are going to observe the new behaviour and may
further extend our own config and/or propose changes to the shared
config.

Note: `rangeStrategy` is already set to `bump` in the shared config, so
I've removed it from here.

Signed-off-by: schogges <moritz.fleck@konghq.com>