chore(deps): remove vitepress (#4426)
We haven't used out vitepress setup for quite a while, pretty much since
we removed the whyframe dependencies.

I always wanted to revisit this as its super useful to have this sort of
stuff.

But for the moment at least, this removes a dependency we don't use and
removes a bunch of code that didn't work and we weren't supporting.

If we ever want it back we can always pull it back out of git.

Signed-off-by: John Cowen <john.cowen@konghq.com>