feature(kuma-http-api): add custom overlayer that can deref during overlay (#4420)
Adds a smol wrapper around https://github.com/thim81/openapi-format
(which is the overlay lib/CLI we are using) to allow us to dereference
`$ref`s _before_ the overlay happens.

I decided to add our own custom CLI instead of composing CLI calls using
`make` because I wanted to avoid making lots of intermediate files
though the pipeline. This way you have one overlay source file and a
single generated OpenAPI spec.

I've made some inline notes to help review, but also note there are a
few formatting changes just due to the multiple bits of
serializing/deserializing happening as the YAML goes through the
pipeline.

~Lastly, we can't really merge this as yet as we need the upstream issue
resolved - worst case scenario we are going to have to patch or fork the
upstream lib - or potentially find another overlay tool/lib that we can
use. Leaving in Draft for the moment, but a totally open for review and
potential informal approval at this stage 👍~

Upstream bugfix is in, I've added a couple of commits on the end of here
for the upgrade of `openapi-format` and then one more `make build` -
[packages/kuma-http-api/dist/components/schemas/MeshAccessLogItem_Put_RequestBody.yaml](https://github.com/kumahq/kuma-gui/pull/4420/changes#diff-9def993296fca062098e47f80d079e0c65813832713efcbfc6fae4157461f549)
is now how we want it 🎉

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>
Signed-off-by: John Cowen <johncowen@users.noreply.github.com>
Co-authored-by: Mo <24444046+schogges@users.noreply.github.com>