chore: refactors CSS underlining (#4425)
Our CSS hover underlining is a little overzealous. We've discussed
moving this into XLayout and making it a little more targeted so it
doesn't put underlines on things we don't want it to by default.

I had to tweak some of the CSS in the PolicyTypeListView also so I
figured I may as well pick up
https://github.com/kumahq/kuma-gui/issues/3093 while I was there.

I also "finally" added a `variant` property to `XLayout` and deprecated
`type`. Moving forwards we should try to use `y-stack`, `x-stack`,
whereas `separated` should only be used for laying out inline lists of
words (where you would commonly use a comma to separate)

Closes https://github.com/kumahq/kuma-gui/issues/3093

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>
Signed-off-by: John Cowen <johncowen@users.noreply.github.com>
Co-authored-by: Mo <24444046+schogges@users.noreply.github.com>