chore(config): create renovate base config (#4428)
Creates a new shareable renovate base config to not require setting same
options in multiple places.

By adopting the base config
`Kong/public-shared-renovate//base/javascript` we've received grouping
of all patches and minors. This is something we'd like to avoid so I've
added the same setting but set `enabled: false`.

Signed-off-by: schogges <moritz.fleck@konghq.com>