fix(mid,mtrust): use generic kri endpoint to fetch resource (#4451)
To fetch a `MeshIdentity` we can use the generic `/_kri/{kri}` endpoint.
This fixes the issue described in
https://github.com/kumahq/kuma-gui/issues/4445.

Additionally I found a bug in the `MeshTrust` listing that occurs when
`spec.origin` is empty. I used this
[example](https://kuma.io/docs/2.12.x/policies/meshtrust/#allow-encrypted-traffic-from-different-trust-domain-to-the-mesh)
to add a `MeshTrust`. I fixed the issue by rendering conditionally.

Unfortunately the same issue described in
https://github.com/kumahq/kuma-gui/issues/4445 appears for `MeshTrust`.
~I've tried the same fix by using the generic `/_kri/{kri}` endpoint,
but the API returns with 404. Need to discuss it further and push a
follow-up for this separately.~ The `MeshTrust` resource I've used was
meant for multizone, but I had a single zone setup. After using a
slightly different config, the `/_kri/{kri}` endpoint also worked for
the `mtrust` kri.

Closes https://github.com/kumahq/kuma-gui/issues/4445

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>