feat(dataplanes): with spire/spiffe conditionally show cert expiration (#4427) (#4454)
>With SPIFFE/SPIRE we get the information about certificate expiration
as part of the envoy stats. There are entries for each service, but
since it's one `MeshIdentity` the expiration time of each entry is
always the same i.e. from a local deployment
>
>```

>cluster.kri_msvc_default_default_kuma-demo_demo-app-v1_5050.ssl.certificate.spiffe://default.local->zone.mesh.local/ns/kuma-demo/sa/default.expiration_unix_time_seconds:
1765831281

>cluster.kri_msvc_default_default_kuma-demo_demo-app-v2_5050.ssl.certificate.spiffe://default.local->zone.mesh.local/ns/kuma-demo/sa/default.expiration_unix_time_seconds:
1765831281

>cluster.kri_msvc_default_default_kuma-demo_demo-app_5050.ssl.certificate.spiffe://default.local->zone.mesh.local/ns/kuma-demo/sa/default.expiration_unix_time_seconds:
1765831281

>cluster.kri_msvc_default_default_kuma-demo_kv_5050.ssl.certificate.spiffe://default.local-zone.mesh.local/ns/kuma->demo/sa/default.expiration_unix_time_seconds:
1765831281

>listener.self_inbound_dp_5050.ssl.certificate.spiffe://default.local-zone.mesh.local/ns/kuma->demo/sa/default.expiration_unix_time_seconds:
1765831281
>```
>
>To get one of the entries I've added a `findDeep` utility function that
recursively tries to find a key within an object. If that key is found
it returns the value. This is needed because the parsing of the stats
returns an arbitrary JSON blob and the location of a spiffe related
entry of `expiration_unix_time_seconds` is unknown. Since there might be
multiple entries that have
`expiration_unix_time_seconds` where some of them are related to SPIFFE
and some are not, I've added a nested `findDeep` to first find an entry
by RegExp `spiffe://` and then `expiration_unix_time_seconds`. To pass
the `expiration_unix_time_seconds` to the view I've added a top level
`$meta` field that stores the information. If the value is not found
(`undefined`) the information remains hidden. Also note that
`expiration_unix_time_seconds` is seconds, so I'm making sure that I
multiply by `1_000` to turn it into milliseconds.

(cherry picked from commit f07a71c17c02d611f0ef07075433a4bc766f1206)

Signed-off-by: schogges <moritz.fleck@konghq.com>