fix(deps): actually remove vitepress (#4456)
I somehow missed this when I did
https://github.com/kumahq/kuma-gui/pull/4426 😅

Interesting removes a version of shiki/core that was a semi-dupe (the
removed version is an older version that what we use int he app). The
version we use seems to be inside of shiki's `node_modules`. I guess
this will be hoisted when we next do a package lock upgrade./maintenance
run.

Signed-off-by: John Cowen <john.cowen@konghq.com>